fixed addons route
